requirements.txt
matplotlib==3.5.2
matplotlib-inline==0.1.3
numpy==1.22.3
onnx==1.12.0
opencv-python==4.5.5.64
packaging==21.3
pandas==1.4.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
pipreqs==0.4.11
prompt-toolkit==3.0.30
protobuf==3.20.1
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.12.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytorch-toolbelt==0.5.0
pytz==2022.1
PyWavelets==1.3.0
PyYAML==6.0
rsa==4.8
scikit-image==0.19.2
scikit-learn==1.0.2
scipy==1.8.0
seaborn==0.11.2
stack-data==0.3.0
tensorboard==2.9.0
threadpoolctl==3.1.0
tifffile==2022.5.4
torch==1.11.0
torch-summary==1.4.5
torchaudio==0.11.0
torchvision==0.12.0
tqdm==4.64.0
traitlets==5.3.0
wcwidth==0.2.5
Werkzeug==2.1.2
yacs==0.1.8
yarg==0.1.9
zipp==3.8.0
